requirements.txt
fastapi>=0.95.0
openai-agents>=0.0.4
pydantic>=1.10.7
requests>=2.28.2
ruff>=0.11.0
uvicorn>=0.21.1
